Main to Staging

+++ app/Http/Controllers/Api/CertificationRequestController.php
@@ -218,7 +218,11 @@ class CertificationRequestController extends Controller
         if ($certificationRequest->status === 'paid') {
             return ApiResponse::error([], 'Cannot update a paid certification request', 403);
         }
-
+        // You can only delete rejected certification request with no initial approval and no membership payments
+        // with no membership payments
+        if ($certificationRequest->status !== 'rejected'  && $data['status'] == 'rejected' && $certificationRequest->membership && $certificationRequest->membership->membershipPayments()->exists()) {
+            return ApiResponse::error([], 'you can only reject certification request with no initial approval and no membership payments', 403);
+        }
         // approved can't be change
         if ($certificationRequest->status === 'approved' && $data['status'] !== 'approved') {
             return ApiResponse::error([], 'Cannot update an approved certification request', 403);
@@ -320,8 +324,19 @@ class CertificationRequestController extends Controller
             // Begin a database transaction
             DB::beginTransaction();
 
+             // Check if the certification request is paid or approved, if yes return error
+             if ($certificationRequest->status === 'paid' || $certificationRequest->status === 'approved') {
+                return ApiResponse::error([], "Can't delete paid or approved certification request", 403);
+            }
+
+            // You can only delete rejected certification request with no initial approval and no membership payments
+            // with no membership payments
+            if ($certificationRequest->status !== 'rejected' && !$certificationRequest->membership && !$certificationRequest->membership->membershipPayments()->exists()) {
+                return ApiResponse::error([], 'you can only delete rejected certification request with no initial approval and no membership payments', 403);
+            }
+
             if ($certificationRequest->status !== 'rejected' || $certificationRequest->membership) {
-                return ApiResponse::error([], 'you can only delete rejected certification request', 403);
+                return ApiResponse::error([], 'you can only delete rejected certification request with no initial approval', 403);
             }
             // Delete the certification request
             $certificationRequest->delete();
